requirements.txt
pandas
numpy
scikit-learn
lightgbm
xgboost
catboost
joblib
fastapi
httpx
uvicorn[standard]
pyarrow
mlflow
optuna
evidently
pyyaml
pytest
flake8
